+++ src/main/java/com/example/askme/api/controller/token/controller/TokenController.java
@@ -1,8 +1,8 @@
 package com.example.askme.api.controller.token.controller;
 
 import com.example.askme.api.controller.token.dto.AccessTokenResponseDto;
-import com.example.askme.api.service.token.serivce.TokenService;
 import com.example.askme.common.ResultResponse;
+import com.example.askme.common.jwt.TokenManager;
 import com.example.askme.common.util.AuthorizationHeaderUtils;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api")
 public class TokenController {
 
-    private final TokenService tokenService;
+    private final TokenManager tokenManager;
 
     @PostMapping("/access-token/issue")
     public ResultResponse<AccessTokenResponseDto> createAccessToken(HttpServletRequest httpServletRequest) {
@@ -23,7 +23,7 @@ public class TokenController {
         AuthorizationHeaderUtils.validateAuthorizationHeader(authorizationHeader);
 
         String refreshToken = authorizationHeader.split(" ")[1];
-        AccessTokenResponseDto accessTokenResponseDto = tokenService.createAccessTokenByRefreshToken(refreshToken);
+        AccessTokenResponseDto accessTokenResponseDto = tokenManager.createAccessTokenByRefreshToken(refreshToken);
         return ResultResponse.success(accessTokenResponseDto);
     }
 }
